requirements.txt
google-cloud-bigquery==3.27.*
python-dotenv==1.1
google-adk>=0.1
